Modified the src files in order for app to build successfully. Updated the vitis build script. Updated the README.
The app builds and runs; logs show the tests pass but the data is empty. With the new SDT structure in Vitis 2024, need to see if the GIC is being configured properly. Maybe the buffer is being filled appropriately.

+++ top/z7_packet_redirection/sw/src/emacps_header.h
@@ -50,7 +50,7 @@
 
 int EmacPsDmaIntrExample(XScuGic *IntcInstancePtr,
 			  XEmacPs *EmacPsInstancePtr,
-			  u16 EmacPsDeviceId, u16 EmacPsIntrId);
+			  u32 EmacPsDeviceId, u16 EmacPsIntrId);
 
 #endif
